django_admin/app/requirements.txt
django==3.2
flake8==4.0
psycopg==3.1.10
python-dotenv==1.0.0
django_split_settings==1.2.0
psycopg2-binary==2.9
django_debug_toolbar==3.4.0
uwsgi
djangorestframework==3.14.0
